fixing the rollback bug.
Essentially, we use the cachetable to keep the undo logs in the memory by saving the
undo entries into the rollback_log_nodes. When we have a large sequential inserts within
one txn, the rollback_log_node will grow and then spill out when it is beyond block_size
and there is a chance that undo log ends up being written to the disk upon checkpoint or eviction.

The complication happens when a child txn commits so that all of its undo logs gets merged into that of its parent
txn. This is done through swinging the log entry list from the rollback_log_node of the
child txn to that of the parent txn. But the impl had forgotten to dirty the parent rollback_log_node...
without any assertions, the bug is very hard to manifest because the node is considered clean and
gets evicted without flushing when under the memory presure, which means the loss of the log entries, but
no warning or failures.